Cleanup code.
