--- PX.Survey.Ext/DAC/TemplateFilter.cs
@@ -1,17 +1,0 @@
-﻿using PX.Data;
-using PX.Data.BQL;
-
-namespace PX.Survey.Ext {
-
-    [PXCacheName("TemplateFilter")]
-    public class TemplateFilter : IBqlTable {
-
-        #region TemplateType
-        public abstract class templateType : BqlString.Field<templateType> { }
-        [PXString(2, IsFixed = true)]
-        [SUComponentType.DetailList]
-        [PXUIField(DisplayName = "Template Type")]
-        public virtual string TemplateType { get; set; }
-        #endregion
-    }
-}
