+++ PX.Survey.Ext/GraphExt/CRActivityMaintExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class CRActivityMaintExt : AbstractSurveyHandlerExt<Objects.EP.CRActivityMaint, Objects.CR.CRActivity> { }
+    public class CRActivityMaintExt : AbstractSurveyHandlerExt<Objects.EP.CRActivityMaint, Objects.CR.CRActivity> {
+        public static bool IsActive() => true;
+    }
 }
