+++ PX.Survey.Ext/GraphExt/CRCaseMaintExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class CRCaseMaintExt : AbstractSurveyHandlerExt<Objects.CR.CRCaseMaint, Objects.CR.CRCase> { }
+    public class CRCaseMaintExt : AbstractSurveyHandlerExt<Objects.CR.CRCaseMaint, Objects.CR.CRCase> {
+        public static bool IsActive() => true;
+    }
 }
