+++ PX.Survey.Ext/GraphExt/CRTaskMaintExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class CRTaskMaintExt : AbstractSurveyHandlerExt<Objects.CR.CRTaskMaint, Objects.CR.CRActivity> { }
+    public class CRTaskMaintExt : AbstractSurveyHandlerExt<Objects.CR.CRTaskMaint, Objects.CR.CRActivity> {
+        public static bool IsActive() => true;
+    }
 }
