+++ PX.Survey.Ext/GraphExt/CampaignMaintExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class CampaignMaintExt : AbstractSurveyHandlerExt<Objects.CR.CampaignMaint, Objects.CR.CRCampaign> { }
+    public class CampaignMaintExt : AbstractSurveyHandlerExt<Objects.CR.CampaignMaint, Objects.CR.CRCampaign> {
+        public static bool IsActive() => true;
+    }
 }
