+++ PX.Survey.Ext/GraphExt/EPEventMaintExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class EPEventMaintExt : AbstractSurveyHandlerExt<Objects.EP.EPEventMaint, Objects.CR.CRActivity> { }
+    public class EPEventMaintExt : AbstractSurveyHandlerExt<Objects.EP.EPEventMaint, Objects.CR.CRActivity> {
+        public static bool IsActive() => true;
+    }
 }
