+++ PX.Survey.Ext/GraphExt/LeadMaintExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class LeadMaintExt : AbstractSurveyHandlerExt<Objects.CR.LeadMaint, Objects.CR.CRLead> { }
+    public class LeadMaintExt : AbstractSurveyHandlerExt<Objects.CR.LeadMaint, Objects.CR.CRLead> {
+        public static bool IsActive() => true;
+    }
 }
