+++ PX.Survey.Ext/GraphExt/OpportunityMaintExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class OpportunityMaintExt : AbstractSurveyHandlerExt<Objects.CR.OpportunityMaint, Objects.CR.CROpportunity> { }
+    public class OpportunityMaintExt : AbstractSurveyHandlerExt<Objects.CR.OpportunityMaint, Objects.CR.CROpportunity> {
+        public static bool IsActive() => true;
+    }
 }
