+++ PX.Survey.Ext/GraphExt/PMQuoteMaintExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class PMQuoteMaintExt : AbstractSurveyHandlerExt<Objects.PM.PMQuoteMaint, Objects.PM.PMQuote> { }
+    public class PMQuoteMaintExt : AbstractSurveyHandlerExt<Objects.PM.PMQuoteMaint, Objects.PM.PMQuote> {
+        public static bool IsActive() => true;
+    }
 }
