+++ PX.Survey.Ext/GraphExt/POOrderEntryExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class POOrderEntryExt : AbstractSurveyHandlerExt<Objects.PO.POOrderEntry, Objects.PO.POOrder> { }
+    public class POOrderEntryExt : AbstractSurveyHandlerExt<Objects.PO.POOrderEntry, Objects.PO.POOrder> {
+        public static bool IsActive() => true;
+    }
 }
