+++ PX.Survey.Ext/GraphExt/QuoteMaintExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class QuoteMaintExt : AbstractSurveyHandlerExt<Objects.CR.QuoteMaint, Objects.CR.CRQuote> { }
+    public class QuoteMaintExt : AbstractSurveyHandlerExt<Objects.CR.QuoteMaint, Objects.CR.CRQuote> {
+        public static bool IsActive() => true;
+    }
 }
