+++ PX.Survey.Ext/GraphExt/SOOrderEntryExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class SOOrderEntryExt : AbstractSurveyHandlerExt<Objects.SO.SOOrderEntry, Objects.SO.SOOrder> { }
+    public class SOOrderEntryExt : AbstractSurveyHandlerExt<Objects.SO.SOOrderEntry, Objects.SO.SOOrder> {
+        public static bool IsActive() => true;
+    }
 }
