+++ PX.Survey.Ext/GraphExt/SOShipmentEntryExt.cs
@@ -1,3 +1,5 @@
 ﻿namespace PX.Survey.Ext {
-    public class SOShipmentEntryExt : AbstractSurveyHandlerExt<Objects.SO.SOShipmentEntry, Objects.SO.SOShipment> { }
+    public class SOShipmentEntryExt : AbstractSurveyHandlerExt<Objects.SO.SOShipmentEntry, Objects.SO.SOShipment> {
+        public static bool IsActive() => true;
+    }
 }
