+++ PX.Survey.Ext/SurveyComponentMaint.cs
@@ -4,10 +4,8 @@ namespace PX.Survey.Ext {
     public class SurveyComponentMaint : PXGraph<SurveyComponentMaint, SurveyComponent> {
 
         public PXSelect<SurveyComponent> SUComponent;
-
         public PXSelect<SurveyComponent, Where<SurveyComponent.componentID, Equal<Current<SurveyComponent.componentID>>>> CurrentSUComponent;
 
-        public SurveyComponentMaint() {
-        }
+        public SurveyComponentMaint() {}
     }
 }
