+++ PX.Survey.Ext/SurveyMaint.cs
@@ -15,7 +15,6 @@ using System.Web;
 namespace PX.Survey.Ext {
     public class SurveyMaint : PXGraph<SurveyMaint, Survey> {
 
-        //[PXCopyPasteHiddenFields(typeof(Survey.surveyID))]
         public SelectFrom<Survey>.View Survey;
         public PXSelect<Survey, Where<Survey.surveyID, Equal<Current<Survey.surveyID>>>> CurrentSurvey;
 
@@ -65,7 +64,7 @@ namespace PX.Survey.Ext {
                 GroupBy<SurveyAnswer.value,
                 Count<SurveyAnswer.lineNbr>>>>>,
             OrderBy<Asc<SurveyDetail.pageNbr, Asc<SurveyDetail.questionNbr, Desc<SurveyAnswer.value>>>>> AnswerSummary;
-        //        public PXSelect<SurveyAnswerSummary, Where<SurveyAnswerSummary.surveyID, Equal<Current<Survey.surveyID>>>> AnswerSummary;
+//        public PXSelect<SurveyAnswerSummary, Where<SurveyAnswerSummary.surveyID, Equal<Current<Survey.surveyID>>>> AnswerSummary;
 
         [PXCopyPasteHiddenView]
         public PXSelectJoin<SurveyAnswer,
@@ -258,11 +257,11 @@ namespace PX.Survey.Ext {
             }
         }
 
-        public PXFilter<TemplateFilter> templatefilter;
+        public PXFilter<ComponentFilter> templatefilter;
 
         [PXFilterable]
         [PXCopyPasteHiddenView]
-        public TemplateLookup<TemplateSelected, TemplateFilter> templates;
+        public ComponentLookup<ComponentSelected, ComponentFilter> templates;
 
         public PXAction<Survey> addTemplates;
         [PXUIField(DisplayName = "Add Pages", MapEnableRights = PXCacheRights.Select, MapViewRights = PXCacheRights.Select)]
@@ -283,11 +282,11 @@ namespace PX.Survey.Ext {
         [PXLookupButton]
         public virtual IEnumerable AddSelectedTemplate(PXAdapter adapter) {
             Details.Cache.ForceExceptionHandling = true;
-            foreach (TemplateSelected template in templates.Cache.Cached) {
+            foreach (ComponentSelected template in templates.Cache.Cached) {
                 if (template.Selected == true) {
                     var surveyDetail = new SurveyDetail();
                     surveyDetail.SurveyID = Survey.Current.SurveyID;
-                    surveyDetail.ComponentType = template.TemplateType;
+                    surveyDetail.ComponentType = template.ComponentType;
                     surveyDetail.Description = template.Description;
                     surveyDetail.ComponentID = template.ComponentID;
                     Details.Update(surveyDetail);
@@ -497,15 +496,6 @@ namespace PX.Survey.Ext {
         [PXUIField(DisplayName = "Load Collectors", MapEnableRights = PXCacheRights.Select, MapViewRights = PXCacheRights.Select)]
         [PXLookupButton]
         public virtual IEnumerable LoadCollectors(PXAdapter adapter) {
-            //if (Survey.Current != null) {
-            //    Save.Press();
-            //    foreach (var user in Users.Select()) {
-            //        var collector = DoUpsertCollector(Survey.Current, user, null);
-            //    }
-            //    Actions.PressSave();
-            //    Collectors.View.RequestRefresh();
-            //}
-            //return adapter.Get();
             Save.Press();
             var list = adapter.Get<Survey>().ToList();
             PXLongOperation.StartOperation(this, delegate () {
@@ -667,7 +657,6 @@ namespace PX.Survey.Ext {
                 }
                 Collectors.Current = collector;
                 var token = collector.Token;
-                //var (_, user, _) = SurveyUtils.GetSurveyAndUser(this, token);
                 var collectorDatas = CollectorDataRecords.Select().FirstTableItems;
                 foreach (var collData in collectorDatas) {
                     if (collData.Status == CollectorDataStatus.Processed ||
@@ -791,24 +780,6 @@ namespace PX.Survey.Ext {
             }
         }
 
-        //protected virtual void _(Events.RowSelecting<Survey> e) {
-        //    Survey row = e.Row;
-        //    if (row == null) { return; }
-        //    using (new PXConnectionScope()) {
-        //        var collectors = SelectFrom<SurveyCollector>.Where<SurveyCollector.surveyID.IsEqual<@P.AsString>>.
-        //                                View.SelectWindowed(this, 0, 1, row.SurveyID);
-        //        row.IsSurveyInUse = collectors.Any();
-        //    }
-        //}
-
-        protected virtual void _(Events.RowPersisted<Survey> e) {
-            var row = e.Row;
-            if (row == null) { return; }
-            recipients.Cache.Clear();
-            recipients.View.RequestRefresh();
-        }
-
-
         protected virtual void _(Events.RowSelected<Survey> e) {
             var row = e.Row;
             if (row == null) { return; }
@@ -819,7 +790,6 @@ namespace PX.Survey.Ext {
             generateSample.SetEnabled(hasPages);
             PXUIFieldAttribute.SetEnabled<Survey.target>(e.Cache, row, !lockedSurvey);
             PXUIFieldAttribute.SetEnabled<Survey.layout>(e.Cache, row, !lockedSurvey);
-            //PXUIFieldAttribute.SetEnabled<Survey.entityType>(e.Cache, row, unlockedSurvey);
             var isAnon = row.Target == SurveyTarget.Anonymous;
             var isEntity = row.Target == SurveyTarget.Entity;
             PXUIFieldAttribute.SetEnabled<Survey.allowAnonymous>(e.Cache, row, !isAnon);
@@ -872,25 +842,6 @@ namespace PX.Survey.Ext {
             var url = generator.GetUrl(row, row.SurveyID, null);
             e.ReturnValue = url;
         }
-
-        //protected virtual void _(Events.RowPersisted<Survey> e) {
-        //    var row = e.Row;
-        //    if (row == null)
-        //        return;
-        //    if (e.TranStatus == PXTranStatus.Completed) {
-        //        //DoResetPageNumbers(row);
-        //        //DoResetQuestionNumbers(row);
-        //    }
-        //}
-
-        //protected virtual void _(Events.RowPersisted<SurveyDetail> e) {
-        //    var row = e.Row;
-        //    if (row == null)
-        //        return;
-        //    if (e.TranStatus == PXTranStatus.Completed) {
-        //        DoInsertMissing(row);
-        //    }
-        //}
 
         protected virtual void _(Events.RowPersisting<SurveyCollector> e) {
             var row = e.Row;
@@ -1163,112 +1114,11 @@ namespace PX.Survey.Ext {
         }
 
         public void _(Events.RowSelected<SurveyAnswer> e) {
-            //var survey = Survey.Current;
             var hasAnswers = Answers.Select().Any();
             reProcessAnswers.SetEnabled(hasAnswers);
             var hasUnProcAnswers = UnprocessedCollectorDataRecords.Select().Any();
             processAnswers.SetEnabled(hasUnProcAnswers);
         }
-
-        //protected virtual void _(Events.FieldSelecting<SurveyAnswer, SurveyAnswer.value> e) {
-        //    var row = e.Row;
-        //    if (row == null || row.AttributeID == null) {
-        //        return;
-        //    }
-        //    int? controlType;
-        //    int num1;
-        //    List<CRAttribute.AttributeValue> values;
-        //    object value;
-        //    CRAttribute.Attribute item = CRAttribute.Attributes[row.AttributeID];
-        //    if (item != null) {
-        //        values = item.Values;
-        //    } else {
-        //        values = null;
-        //    }
-        //    List<CRAttribute.AttributeValue> attributeValues = values;
-        //    bool? isRequired = true;// row.IsRequired;
-        //    int num2 = (isRequired.GetValueOrDefault() & isRequired.HasValue ? 1 : -1);
-        //    if (attributeValues != null && attributeValues.Count > 0) {
-        //        List<string> strs = new List<string>();
-        //        List<string> strs1 = new List<string>();
-        //        foreach (CRAttribute.AttributeValue attributeValue in attributeValues) {
-        //            if (attributeValue.Disabled && row.Value != attributeValue.ValueID) {
-        //                continue;
-        //            }
-        //            strs.Add(attributeValue.ValueID);
-        //            strs1.Add(attributeValue.Description);
-        //        }
-        //        e.ReturnState = PXStringState.CreateInstance(e.ReturnState, new int?(10), new bool?(true), typeof(SurveyAnswer.value).Name, new bool?(false), new int?(num2), item.EntryMask, strs.ToArray(), strs1.ToArray(), new bool?(true), null, null);
-        //        controlType = item.ControlType;
-        //        if (controlType.GetValueOrDefault() == 6 & controlType.HasValue) {
-        //            ((PXStringState)e.ReturnState).MultiSelect = true;
-        //            if (e.Cache.Graph.IsContractBasedAPI) {
-        //                string returnValue = e.ReturnValue as string;
-        //                if (returnValue != null) {
-        //                    e.ReturnValue = string.Join(", ", returnValue.Split(new char[] { ',' }).Select<string, string>((string i) => {
-        //                        int num = strs.IndexOf(i.Trim());
-        //                        if (num < 0) {
-        //                            return i;
-        //                        }
-        //                        return strs1[num];
-        //                    }));
-        //                }
-        //            }
-        //        }
-        //    } else if (item != null) {
-        //        controlType = item.ControlType;
-        //        if (!(controlType.GetValueOrDefault() == 4 & controlType.HasValue)) {
-        //            controlType = item.ControlType;
-        //            if (!(controlType.GetValueOrDefault() == 5 & controlType.HasValue)) {
-        //                PXStringState stateExt = e.Cache.GetStateExt<SurveyAnswer.value>(null) as PXStringState;
-        //                //PXFieldSelectingEventArgs pXFieldSelectingEventArg = e;
-        //                object returnState = e.ReturnState;
-        //                PXStringState pXStringState = stateExt;
-        //                isRequired = null;
-        //                e.ReturnState = PXStringState.CreateInstance(returnState, new int?(pXStringState.With((PXStringState _) => _.Length)), isRequired, typeof(SurveyAnswer.value).Name, new bool?(false), new int?(num2), item.EntryMask, null, null, new bool?(true), null, null);
-        //            } else {
-        //                DateTime? nullable = null;
-        //                DateTime? nullable1 = nullable;
-        //                nullable = null;
-        //                e.ReturnState = PXDateState.CreateInstance(e.ReturnState, typeof(SurveyAnswer.value).Name, new bool?(false), new int?(num2), item.EntryMask, item.EntryMask, nullable1, nullable);
-        //            }
-        //        } else {
-        //            object obj = e.ReturnState;
-        //            Type type = typeof(bool);
-        //            bool? nullable2 = new bool?(false);
-        //            bool? nullable3 = new bool?(false);
-        //            int? nullable4 = new int?(num2);
-        //            controlType = null;
-        //            int? nullable5 = controlType;
-        //            controlType = null;
-        //            isRequired = null;
-        //            e.ReturnState = PXFieldState.CreateInstance(obj, type, nullable2, nullable3, nullable4, nullable5, controlType, false, typeof(SurveyAnswer.value).Name, null, null, null, PXErrorLevel.Undefined, new bool?(true), new bool?(true), isRequired, PXUIVisibility.Visible, null, null, null);
-        //            if (e.ReturnValue is string && int.TryParse((string)e.ReturnValue, NumberStyles.Integer, CultureInfo.InvariantCulture, out num1)) {
-        //                e.ReturnValue = Convert.ToBoolean(num1);
-        //            }
-        //        }
-        //    }
-        //    if (e.ReturnState is PXFieldState) {
-        //        PXFieldState errorText = (PXFieldState)e.ReturnState;
-        //        IPXInterfaceField pXInterfaceField = e.Cache.GetAttributes(row, typeof(SurveyAnswer.value).Name).OfType<IPXInterfaceField>().FirstOrDefault<IPXInterfaceField>();
-        //        if (pXInterfaceField != null && pXInterfaceField.ErrorLevel != PXErrorLevel.Undefined && !string.IsNullOrEmpty(pXInterfaceField.ErrorText)) {
-        //            errorText.Error = pXInterfaceField.ErrorText;
-        //            errorText.ErrorLevel = pXInterfaceField.ErrorLevel;
-        //        }
-        //        //PXFieldState valueExt = sender.GetValueExt<SurveyAnswer.attributeCategory>(row) as PXFieldState;
-        //        //if (valueExt != null) {
-        //        //    value = valueExt.Value;
-        //        //} else {
-        //        value = null;
-        //        //}
-        //        errorText.Enabled = (string)value != "V";
-        //        if (IsContractBasedAPI) {
-        //            errorText.ErrorLevel = PXErrorLevel.Undefined;
-        //            errorText.Error = null;
-        //            e.Cancel = true;
-        //        }
-        //    }
-        //}
 
         protected virtual void _(Events.FieldUpdating<SurveyAnswer, SurveyAnswer.value> e) {
             var row = e.Row;
@@ -1276,8 +1126,8 @@ namespace PX.Survey.Ext {
                 return;
             }
             int? controlType;
-            bool flag;
-            int num;
+            //bool flag;
+            //int num;
             DateTime dateTime;
             CRAttribute.Attribute item = CRAttribute.Attributes[row.AttributeID];
             if (item == null) {
@@ -1311,8 +1161,8 @@ namespace PX.Survey.Ext {
                     case 3:
                         break;
                     case 4:
-                        if (bool.TryParse(str, out flag)) {
-                            num = Convert.ToInt32(flag);
+                        if (bool.TryParse(str, out var flag)) {
+                            var num = Convert.ToInt32(flag);
                             e.NewValue = num.ToString(CultureInfo.InvariantCulture);
                             return;
                         }
@@ -1336,7 +1186,7 @@ namespace PX.Survey.Ext {
                             break;
                         }
                         string[] strArrays = str.Split(new char[] { ',' });
-                        for (num = 0; num < (int)strArrays.Length; num++) {
+                        for (var num = 0; num < (int)strArrays.Length; num++) {
                             string str1 = strArrays[num];
                             if (item.Values.Find((CRAttribute.AttributeValue a) => string.Equals(a.ValueID, str1, StringComparison.OrdinalIgnoreCase)) == null) {
                                 e.Cache.RaiseExceptionHandling<SurveyAnswer.value>(row, str, new PXSetPropertyException("One of the specified values is not valid for the {0} attribute that has the Multi Select Combo type. Note that the Multi Select Combo type supports identifiers and does not support descriptions.", new object[] { row.AttributeID }));
